pyproject.toml
[project]
name = "book-club-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord>=2.3.2",
    "python-dotenv>=1.2.2",
]
